Refactor to single db.
Add chat table.
Add channel table.
Add user table.
Add normal_channel table.
Add normal_user table.
Add open_channel table.
Add open_user table.
