+++ crates/kiwi-talk-db/src/channel/normal/mod.rs
@@ -1,0 +1,13 @@
+pub mod model;
+
+use rusqlite::Connection;
+
+#[derive(Debug, Clone, Copy)]
+pub struct NormalChannelEntry<'a>(pub &'a Connection);
+
+impl<'a> NormalChannelEntry<'a> {}
+
+#[derive(Debug, Clone, Copy)]
+pub struct NormalUserEntry<'a>(pub &'a Connection);
+
+impl<'a> NormalUserEntry<'a> {}
