+++ crates/kiwi-talk-db/src/channel/normal/model.rs
@@ -1,0 +1,11 @@
+#[derive(Debug, PartialEq, Eq)]
+pub struct NormalChannelModel {
+    pub id: i64,
+    pub join_time: i64,
+}
+
+#[derive(Debug, PartialEq, Eq)]
+pub struct NormalUserModel {
+    pub id: i64,
+    
+}
