+++ crates/kiwi-talk-db/src/channel/open/model.rs
@@ -1,0 +1,9 @@
+#[derive(Debug, PartialEq, Eq)]
+pub struct OpenChannelModel {
+    pub id: i64,
+}
+
+#[derive(Debug, PartialEq, Eq)]
+pub struct OpenUserModel {
+    pub id: i64
+}
